+++ src/index.ts
@@ -1,6 +1,5 @@
-export { default as headerAuthentication } from './headerAuthentication';
-export * from './middleware';
-export * from './error';
-export { Rule, default as jwtAuthorization } from './jwtAuthorization';
-export * from './validation';
 export { AppInsightForHttpTrigger, AppInsightForNoNHttpTrigger } from './appInsights/appInsightsWrapper';
+export * from './authentication';
+export * from './error';
+export * from './middleware';
+export * from './validation';